Image encoding fixes and GIMP export of palette data
Fixed some bugs with encoding of images. Should work now properly. Can
only decode RGB or RGBA PNG data for now. Not really planning on fixing
it, it works for my use case. Also added export of the Doom palette data
so that it can be used from within Gimp.